refactor(tracker): split utils.py into utils/ package
Split the 1949-line tracker/utils.py into focused submodules
(resources, reporting, task_execution, run_control, orchestration,
harness_config) behind a re-exporting __init__.py shim so all existing
imports keep working. Function bodies are unchanged (byte-identical).
Taskiq task name for process_benchmark pinned to its original value.
Test monkeypatch targets repointed to the new submodules.
