--- services/tracker/src/tracker/utils.py
@@ -1,1948 +1,0 @@
-import asyncio
-import base64
-import io
-import json
-import time
-import traceback
-from asyncio import Semaphore, gather
-from collections.abc import AsyncGenerator, Buffer, Coroutine
-from contextlib import suppress
-from datetime import datetime
-from enum import Enum
-from functools import cached_property
-from typing import Any, NamedTuple, Sequence, cast
-from uuid import UUID
-from zoneinfo import ZoneInfo
-
-import logfire
-import sentry_sdk
-from benchmark_service import (
-    Sandbox,
-    SandboxNotFoundError,
-    SandboxProvider,
-    SandboxProviderConfig,
-    SandboxQuery,
-    sandbox_provider_config_from_mapping,
-)
-from benchmark_service.client import BenchmarkServiceClient, BenchmarkServiceError, BenchmarkServiceUnauthenticatedError
-from fastapi import HTTPException, Request
-from opentelemetry import trace
-from pydantic import ValidationError
-from sqlalchemy import JSON, literal, tuple_, type_coerce
-from sqlalchemy.orm import selectinload
-from sqlmodel import Session, asc, case, col, desc, func, or_, select, update
-from tenacity import retry as tenacity_retry
-from tenacity import retry_if_exception_type, stop_after_attempt, wait_fixed
-from websockets.exceptions import ConnectionClosedError, InvalidStatus
-
-from tracker._lambda import invoke_lambda, lambda_client
-from tracker.auth import RequestIdentity
-from tracker.aws.cloudwatch_logs import create_benchmark_log_group, write_benchmark_log_event
-from tracker.aws.s3 import (
-    S3_BENCHMARKS_PREFIX,
-    copy_agent_to_benchmark,
-    create_benchmark_url,
-    get_agent_result_s3_key,
-    upload_to_s3,
-)
-from tracker.aws.secrets import fetch_aws_secret, resolve_secrets
-from tracker.config import ENVIRONMENT, broker, create_benchmark_service_url
-from tracker.database.models import (
-    Benchmark,
-    BenchmarkArguments,
-    BenchmarkStatus,
-    DocentReadingStatus,
-    ErrorResult,
-    EvaluationResult,
-    FinalEvaluation,
-    Org,
-    RetryMode,
-    Task,
-    TaskBreakdown,
-    TaskStatus,
-)
-from tracker.database.scoping import scoped_select
-from tracker.database.session import engine
-from tracker.exceptions import OutputArtifactError, SandboxSetupError, TrackerServiceError
-from tracker.logging import get_logger, task_id_var
-from tracker.notifications import NotificationContext, SlackNotifier
-from tracker.observability import elapsed_ms, retry_callback
-from tracker.sandbox import create_sandbox, delete_sandbox, run_agent, upload_agent_artifacts
-from tracker.types import (
-    AverageTaskBreakdown,
-    AWSCredentials,
-    BenchmarkDetails,
-    BenchmarkTableRow,
-    FetchBenchmarkResponse,
-    FetchBenchmarksRequest,
-    FinalViewResponse,
-    HarnessConfig,
-    Order,
-    StartBenchmarkRequest,
-)
-
-logger = get_logger(__name__)
-
-_SANDBOX_CREATION_CAP: int = 10
-_PTY_TASK_RETRY_LIMIT: int = 1
-_RUNNABLE_TASK_STATUSES = [TaskStatus.PENDING, TaskStatus.BUILDING, TaskStatus.IN_PROGRESS, TaskStatus.EVALUATING]
-
-
-def _history_result(created_at: datetime, result: dict[str, Any]) -> dict[str, Any]:
-    return {"created_at": created_at.isoformat(), "result": dict(result)}
-
-
-def _history_error(created_at: datetime, error_message: str) -> dict[str, Any]:
-    return {"created_at": created_at.isoformat(), "error_message": error_message}
-
-
-def fetch_sandbox_provider_config(secret_name: str, aws: AWSCredentials, provider_type: str) -> SandboxProviderConfig:
-    """Resolve sandbox provider config from the selected provider type and secret."""
-    secret = fetch_aws_secret(secret_name, aws)
-    if not isinstance(secret, dict):
-        raise TrackerServiceError("Expected sandbox provider secret to be a JSON object")
-
-    return sandbox_provider_config_from_mapping({**secret, "type": provider_type})
-
-
-def create_benchmark_service_client(
-    url: str,
-    service_headers: dict[str, str] | None = None,
-) -> BenchmarkServiceClient:
-    """Create a BenchmarkServiceClient with benchmark-service headers."""
-    headers: dict[str, str] = {}
-    if service_headers:
-        headers.update(service_headers)
-    return BenchmarkServiceClient(url=url, headers=headers)
-
-
-def create_benchmark_service_client_from_request(request: StartBenchmarkRequest) -> BenchmarkServiceClient:
-    """Create a BenchmarkServiceClient for a start request."""
-    url = request.custom_benchmark_service or create_benchmark_service_url(request.benchmark_name)
-    return create_benchmark_service_client(url, service_headers=request.service_headers)
-
-
-def start_benchmark_request_to_benchmark(request: StartBenchmarkRequest, run_starter: RequestIdentity) -> Benchmark:
-    """Convert a StartBenchmarkRequest to a Benchmark database model."""
-    return Benchmark(
-        org_id=run_starter.org.id,
-        name=request.benchmark_name,
-        label=request.label,
-        custom_benchmark_service=request.custom_benchmark_service,
-        webhook_secret_name=request.webhook_secret_name,
-        webhook_intervals=request.webhook_intervals,
-        arguments=BenchmarkArguments(
-            contract=request.contract,
-            concurrency=request.concurrency,
-            task_ids=request.task_ids,
-            slice_str=request.slice_str,
-            lambda_function=request.lambda_function,
-            dataset=request.dataset,
-            sandbox_provider=request.sandbox_provider,
-            sandbox_provider_secret_name=request.harness_config.sandbox_provider_secret_name,
-        ),
-        started_by_id=run_starter.access_key_id,
-        started_by_email=run_starter.email,
-    )
-
-
-class TrackedTaskStatus(str, Enum):
-    WAITING = "waiting"
-    RUNNING = "running"
-    DONE = "done"
-
-
-class TrackedTask:
-    _coro: Coroutine[Any, Any, Any]
-    _status: str
-    _task: asyncio.Task[Any] | None
-    _org: Org
-
-    def __init__(self, coro: Coroutine[Any, Any, Any], org: Org):
-        self._coro = coro
-        self._org = org
-        self._status = TrackedTaskStatus.WAITING
-        self._task = None
-
-    @property
-    def status(self) -> str:
-        return self._status
-
-    @property
-    def task(self) -> asyncio.Task[Any] | None:
-        return self._task
-
-    async def run(self, semaphore: asyncio.Semaphore, task_row: Task) -> dict[str, dict[str, Any] | None]:
-        async def _wrap_coro():
-            """Need to have a task created even if we are not running the coroutine so that we can cancel it before its running"""
-            async with semaphore:
-                self._status = TrackedTaskStatus.RUNNING
-                return await self._coro
-
-        try:
-            self._task = asyncio.create_task(_wrap_coro())
-            return await self._task
-        except asyncio.CancelledError:
-            logger.warning(f"Task {task_row.task_id} was cancelled")
-            # Need to clean up the coroutine if we cancelled the task
-            self._coro.close()
-
-            # When we cancel we return the task id still so that we can track the task when we create the final evaluation row
-            return {task_row.task_id: None}
-        except Exception as e:
-            error_message = f"Task error was not handled: {str(e)}\n{traceback.format_exc()}"
-            logger.error(error_message)
-            logfire.exception("tracked_task_run failed")
-            sentry_sdk.capture_exception(e)
-            with Session(bind=engine) as session:
-                task = fetch_task_row(task_row.id, session, self._org)
-                commit_task_error(task, session, error_message)
-
-            return {task_row.task_id: None}
-        finally:
-            self._status = TrackedTaskStatus.DONE
-
-
-class TaskMonitor:
-    _benchmark_id: UUID
-    _task_tracking: dict[str, TrackedTask]
-    _notifier: SlackNotifier | None
-    _org: Org
-    _TRACK_INTERVAL: int = 2
-
-    def __init__(
-        self, benchmark_id: UUID, task_tracking: dict[str, TrackedTask], org: Org, notifier: SlackNotifier | None = None
-    ):
-        self._benchmark_id = benchmark_id
-        self._task_tracking = task_tracking
-        self._org = org
-        self._notifier = notifier
-
-    def _fetch_task_row(self, task_id: str) -> Task:
-        with Session(bind=engine) as session:
-            task_row = session.exec(
-                select(Task)
-                .where(Task.task_id == task_id)
-                .where(Task.benchmark == self._benchmark_id)
-                .where(Task.org_id == self._org.id)
-                .limit(1)
-            ).first()
-
-            if not task_row:
-                raise ValueError(f"Task with id {task_id} not found")
-
-            return task_row
-
-    def _validate_task(self, task_id: str) -> bool:
-        """
-        If the task status has been set to stopped we return False to exit the task early.
-
-        Returns:
-            True if the task should continue to be processed, False if the task should be stopped early
-
-        """
-        with Session(bind=engine) as session:
-            benchmark_row = fetch_benchmark_row(self._benchmark_id, session, self._org)
-            task_row = self._fetch_task_row(task_id)
-
-            # If task has been stopped or benchmark has errored we need to exit
-            if task_row.status == TaskStatus.STOPPED or benchmark_row.status == BenchmarkStatus.ERROR:
-                return False
-
-        return True
-
-    async def _check_notifications(self) -> None:
-        """Check notification thresholds using DB task counts."""
-        if not self._notifier:
-            return
-
-        with Session(bind=engine) as session:
-            benchmark_row = fetch_benchmark_row(self._benchmark_id, session, self._org)
-            notification_context = NotificationContext.from_benchmark(benchmark_row, session, self._org)
-            await self._notifier.check_and_notify(notification_context)
-
-    async def track_tasks(self) -> None:
-        """
-        Tracks tasks and cancels them when they are no longer valid.
-        """
-
-        exit_condition_met: bool = False
-
-        while not exit_condition_met and self._task_tracking:
-            tasks_to_check: list[str] = list(self._task_tracking.keys())
-            for task_id in tasks_to_check:
-                task = self._task_tracking[task_id]
-
-                if task.status == TrackedTaskStatus.DONE:
-                    del self._task_tracking[task_id]
-                    continue
-
-                if not self._validate_task(task_id) and task.task is not None and not task.task.done():
-                    task.task.cancel(f"Task {task_id} has been invalidated. Run has been requested to stop")
-
-            await self._check_notifications()
-
-            if not self._task_tracking:
-                exit_condition_met = True
-
-            await asyncio.sleep(self._TRACK_INTERVAL)
-
-
-def fetch_benchmark_row(benchmark_id: UUID, session: Session, org: Org) -> Benchmark:
-    """Fetch benchmark row with org validation. Raises domain errors (not HTTPException) for use in background tasks."""
-    benchmark_row = session.get(Benchmark, benchmark_id)
-    if not benchmark_row:
-        raise ValueError(f"Run with id {benchmark_id} not found")
-    if benchmark_row.org_id != org.id:
-        raise ValueError(f"Run {benchmark_id} does not belong to org {org.id}")
-    return benchmark_row
-
-
-def handle_early_exit(task_row: Task, task_session: Session) -> None:
-    _commit_task_status(task_row, task_session, TaskStatus.STOPPED)
-
-
-def fetch_task_row(task_id: UUID, session: Session, org: Org) -> Task:
-    """Fetch task row with org validation. Raises domain errors (not HTTPException) for use in background tasks."""
-    task_row = session.get(Task, task_id)
-    if not task_row:
-        raise TrackerServiceError(f"Task with id {task_id} not found")
-    if task_row.org_id != org.id:
-        raise TrackerServiceError(f"Task {task_id} does not belong to org {org.id}")
-    return task_row
-
-
-def buffer_logs(
-    log_queue: asyncio.Queue[str], stream_key: str, aws: AWSCredentials, log_group: str, force_flush: bool = False
-) -> None:
-    """
-    Buffers the logs in the queue and waits till they are full before streaming them to CloudWatch.
-    """
-    if not log_queue.full() and not force_flush:
-        return
-
-    messages: list[str] = []
-    while not log_queue.empty():
-        messages.append(log_queue.get_nowait())
-
-    message = "".join(messages)
-    loop = asyncio.get_running_loop()
-    loop.run_in_executor(None, write_benchmark_log_event, stream_key, message, aws, log_group)
-
-
-def save_eval_resume_state(task_row_id: UUID, org: Org, eval_resume_state: dict[str, Any]) -> None:
-    with Session(bind=engine) as session:
-        task = fetch_task_row(task_row_id, session, org)
-        task.eval_resume_state = eval_resume_state
-        session.commit()
-
-
-def _commit_task_status(
-    task: Task,
-    session: Session,
-    to_status: TaskStatus,
-    *,
-    error_message: str | None = None,
-    extra: dict[str, Any] | None = None,
-) -> None:
-    from_status = task.status
-    span_attributes = {
-        "benchmark_id": str(task.benchmark),
-        "task_id": task.task_id,
-        "from_status": from_status.value,
-        "to_status": to_status.value,
-        **(extra or {}),
-    }
-    if error_message is not None:
-        span_attributes["has_error_message"] = True
-
-    with logfire.span("task.status_transition", **span_attributes):  # pyright: ignore[reportArgumentType]
-        task.status = to_status
-        session.add(task)
-        session.commit()
-
-
-def commit_task_status_transition(task_row_id: UUID, session: Session, org: Org, to_status: TaskStatus) -> None:
-    fetch_start = time.monotonic()
-    task = fetch_task_row(task_row_id, session, org)
-    _commit_task_status(task, session, to_status, extra={"fetch_duration_ms": elapsed_ms(fetch_start)})
-
-
-@logfire.instrument("process_task")
-@tenacity_retry(
-    retry=retry_if_exception_type(SandboxSetupError),
-    stop=stop_after_attempt(_PTY_TASK_RETRY_LIMIT + 1),
-    wait=wait_fixed(2),
-    before_sleep=retry_callback("valkyrie.task"),
-    reraise=True,
-)
-async def process_task(
-    task_row: Task,
-    start_benchmark_request: StartBenchmarkRequest,
-    benchmark_service: BenchmarkServiceClient,
-    benchmark_id: UUID,
-    task_id: str,
-    harness_config: HarnessConfig,
-    org: Org,
-    sandbox_provider_config: SandboxProviderConfig,
-    creation_semaphore: Semaphore,
-) -> dict[str, dict[str, Any] | None]:
-    """
-    Processes a task and returns the evaluation result
-
-    NOTE: When we close the sandbox the agent process will be killed and we will instantly go to evaluating,
-    the evaluation will fail since the instance no longer exists. We handle this inside of the exception caught.
-    """
-    task_id_var.set(task_id)
-    sentry_sdk.set_tag("benchmark_name", start_benchmark_request.benchmark_name)
-    sentry_sdk.set_tag("agent_name", start_benchmark_request.contract.name)
-    trace.get_current_span().set_attributes(
-        {
-            "task_id": task_id,
-            "benchmark_id": str(benchmark_id),
-            "benchmark_name": start_benchmark_request.benchmark_name,
-            "agent_name": start_benchmark_request.contract.name,
-        }
-    )
-
-    with Session(bind=engine) as task_session:
-        benchmark_row = fetch_benchmark_row(benchmark_id, task_session, org)
-        task_row = task_session.merge(task_row)
-
-        # If user has requested to stop the benchmark we exit before we process the task
-        if benchmark_row.status == BenchmarkStatus.STOPPING:
-            handle_early_exit(task_row, task_session)
-            return {task_id: None}
-        benchmark_name = benchmark_row.name
-        benchmark_agent_name = benchmark_row.arguments.contract.name
-        benchmark_started_by_email = benchmark_row.started_by_email
-
-    # Setup logging infrastructure before try block so it's always available
-    # Suffix is required to version control streams, never delete between retries
-    stream_suffix = f"{int(task_row.started_at.timestamp() * 1_000_000):x}"
-    stream_key: str = f"{benchmark_id}:{task_id}_{stream_suffix}"
-    log_queue: asyncio.Queue[str] = asyncio.Queue(maxsize=20)
-
-    last_log_time: float = time.monotonic()
-
-    # Collects the logs and dumps them when the queue is full
-    def log_output(data: str) -> None:
-        nonlocal last_log_time
-        last_log_time = time.monotonic()
-        log_queue.put_nowait(data)
-        buffer_logs(log_queue, stream_key, harness_config.aws, harness_config.log_group)
-
-    # Auto flush if process takes a while to produce next log
-    # If a process pauses without producing anymore logs, the logs we have collected get stuck
-    async def auto_flush_logs() -> None:
-        while True:
-            await asyncio.sleep(1)
-            if not log_queue.empty() and time.monotonic() - last_log_time >= 10:
-                buffer_logs(log_queue, stream_key, harness_config.aws, harness_config.log_group, force_flush=True)
-
-    flush_task = asyncio.create_task(auto_flush_logs())
-
-    def on_eval_resume_state(state: dict[str, Any]) -> None:
-        save_eval_resume_state(task_row.id, org, state)
-
-    def task_is_stopped() -> bool:
-        with Session(bind=engine) as task_session:
-            task = fetch_task_row(task_row.id, task_session, org)
-            return task.status == TaskStatus.STOPPED
-
-    try:
-        if task_row.status == TaskStatus.EVALUATING and task_row.eval_resume_state is not None:
-            try:
-                log_output("Resuming evaluation from durable benchmark state\n")
-                resume_eval_start_time = time.perf_counter()
-                # Reset timer to keep the last received message from the benchmarks service accurate
-                last_log_time = time.monotonic()
-                evaluation_result = await benchmark_service.resume_evaluation(
-                    task_row.task_id,
-                    eval_resume_state=task_row.eval_resume_state,
-                    on_message=log_output,
-                    on_eval_resume_state=on_eval_resume_state,
-                    dataset=start_benchmark_request.dataset,
-                )
-                resume_eval_duration = time.perf_counter() - resume_eval_start_time
-                evaluation_result_row = EvaluationResult(
-                    org_id=org.id,
-                    task=task_row.id,
-                    instance_id=None,
-                    result=evaluation_result,
-                    agent_caused_exit_reason=None,
-                )
-
-                with Session(bind=engine) as task_session:
-                    task_session.add(evaluation_result_row)
-                    task_in_session = fetch_task_row(task_row.id, task_session, org)
-                    if task_in_session.task_breakdown:
-                        existing_breakdown = task_session.get(TaskBreakdown, task_in_session.task_breakdown)
-                        assert existing_breakdown is not None
-                        existing_breakdown.evaluation_run_duration = resume_eval_duration
-                    commit_task_status_transition(task_row.id, task_session, org, TaskStatus.FINISHED)
-
-                    return {task_id: evaluation_result_row.result}
-            except Exception as e:
-                with Session(bind=engine) as task_session:
-                    task = fetch_task_row(task_row.id, task_session, org)
-                    if task.status == TaskStatus.STOPPED:
-                        return {task_id: None}
-
-                raise e from e
-
-        task_data = await benchmark_service.retrieve_task(task_id=task_id, dataset=start_benchmark_request.dataset)
-        sandbox_provider = benchmark_service.get_sandbox_provider(sandbox_provider_config)
-
-        # Labels that show up in the UI we can use to filter sandboxes
-        labels = {
-            "Benchmark": start_benchmark_request.benchmark_name,
-            "Id": str(benchmark_id),
-            "Task": task_row.task_id,
-        }
-
-        with Session(bind=engine) as task_session:
-            commit_task_status_transition(task_row.id, task_session, org, TaskStatus.BUILDING)
-
-        identity = {
-            "benchmark_name": benchmark_name,
-            "agent_name": benchmark_agent_name,
-        }
-        if benchmark_started_by_email:
-            identity["email"] = benchmark_started_by_email
-
-        env_vars = {
-            **resolve_secrets(start_benchmark_request.contract.secrets, harness_config.aws),
-            "RUN_ID": str(benchmark_id),
-            "TASK_ID": task_row.task_id,
-            "IDENTITY": json.dumps(identity),
-            # Tags sandbox-internal OTel telemetry with our IDs + environment so traces/logs/metrics
-            # are filterable per benchmark run and separable from other environments sharing the
-            # same Daytona account (sandbox OTLP export is account-level).
-            "DAYTONA_SANDBOX_OTEL_EXTRA_LABELS": (
-                f"benchmark_id={benchmark_id},task_id={task_row.task_id},environment={ENVIRONMENT}"
-            ),
-        }
-
-        # We don't want to track the task until the sandbox is actually created.
-        task_breakdown = TaskBreakdown()
-
-        start_sandbox_build_time = time.perf_counter()
-        async with create_sandbox(
-            provider=sandbox_provider,
-            sandbox_name=task_row.task_id,
-            source=task_data.source,
-            labels=labels,
-            env_vars=env_vars,
-            resources=task_data.resources,
-            creation_semaphore=creation_semaphore,
-        ) as sandbox:
-            task_breakdown.sandbox_build_duration = time.perf_counter() - start_sandbox_build_time
-            start_sandbox_run_time = time.perf_counter()
-
-            try:
-                with Session(bind=engine) as task_session:
-                    commit_task_status_transition(task_row.id, task_session, org, TaskStatus.IN_PROGRESS)
-
-                # Upload the contract to the sandbox after creating and install the dependencies
-                await upload_agent_artifacts(
-                    sandbox,
-                    start_benchmark_request.contract,
-                    str(benchmark_id),
-                    harness_config.aws,
-                    harness_config.s3_bucket,
-                )
-
-                # Reset timer to keep the last received message from the benchmarks service accurate
-                last_log_time = time.monotonic()
-                _ = await benchmark_service.setup_task(
-                    task_row.task_id,
-                    sandbox.id,
-                    on_message=log_output,
-                    dataset=start_benchmark_request.dataset,
-                    sandbox_provider=sandbox_provider_config,
-                )
-
-                # Force flush the logs if anything has been buffered
-                buffer_logs(log_queue, stream_key, harness_config.aws, harness_config.log_group, force_flush=True)
-
-                # Compute the S3 key for the agent's output archive
-                agent_output_s3_key = None
-                if start_benchmark_request.contract.final_output:
-                    agent_output_s3_key = get_agent_result_s3_key(str(benchmark_id), task_id, "agent_output.tar.gz")
-
-                exit_reason, agent_run_time = await run_agent(
-                    sandbox,
-                    start_benchmark_request.contract,
-                    task_data.problem_path,
-                    task_id,
-                    log_output,
-                    task_data.cwd,
-                    aws=harness_config.aws,
-                    s3_bucket=harness_config.s3_bucket,
-                    agent_output_s3_key=agent_output_s3_key,
-                    agent_timeout=task_data.agent_timeout,
-                    benchmark_id=str(benchmark_id),
-                )
-                logger.info(
-                    "agent.run.complete",
-                    extra={
-                        "benchmark_id": str(benchmark_id),
-                        "task_id": task_row.task_id,
-                        "sandbox_id": sandbox.id,
-                        "sandbox_name": sandbox.name,
-                        "exit_reason": exit_reason.value if exit_reason else None,
-                    },
-                )
-
-                task_breakdown.agent_run_duration = agent_run_time
-                with Session(bind=engine) as task_session:
-                    task_session.add(task_breakdown)
-                    task_in_session = fetch_task_row(task_row.id, task_session, org)
-                    task_in_session.task_breakdown = task_breakdown.id
-                    commit_task_status_transition(task_row.id, task_session, org, TaskStatus.EVALUATING)
-
-                # Evaluate the instance
-                evaluation_start_time = time.perf_counter()
-
-                logger.info(
-                    "task.evaluation.start",
-                    extra={
-                        "benchmark_id": str(benchmark_id),
-                        "task_id": task_row.task_id,
-                        "sandbox_id": sandbox.id,
-                        "sandbox_name": sandbox.name,
-                    },
-                )
-                logger.info(f"Evaluating agent {start_benchmark_request.contract.name} in sandbox {sandbox.name}")
-                # Reset timer to keep the last received message from the benchmarks service accurate
-                last_log_time = time.monotonic()
-                evaluation_result = await benchmark_service.evaluate_instance(
-                    task_row.task_id,
-                    sandbox.id,
-                    on_message=log_output,
-                    on_eval_resume_state=on_eval_resume_state,
-                    dataset=start_benchmark_request.dataset,
-                    sandbox_provider=sandbox_provider_config,
-                )
-                task_breakdown.evaluation_run_duration = time.perf_counter() - evaluation_start_time
-
-                task_breakdown.sandbox_run_duration = time.perf_counter() - start_sandbox_run_time
-
-                # Force flush the logs, maybe redundant since we have the one in finally:
-                buffer_logs(log_queue, stream_key, harness_config.aws, harness_config.log_group, force_flush=True)
-
-                # Save the evaluation result to the database with the task row
-                # Record the termination reason if the agent did not exit cleanly (timeout / OS kill)
-                evaluation_result_row = EvaluationResult(
-                    org_id=org.id,
-                    task=task_row.id,
-                    instance_id=sandbox.id,
-                    result=evaluation_result,
-                    agent_caused_exit_reason=exit_reason,
-                )
-
-                with Session(bind=engine) as task_session:
-                    task_session.add(evaluation_result_row)
-                    task_in_session = fetch_task_row(task_row.id, task_session, org)
-                    existing_breakdown = task_session.get(TaskBreakdown, task_in_session.task_breakdown)
-                    if existing_breakdown is None:
-                        raise TrackerServiceError(f"Missing task breakdown for task {task_row.id}")
-                    existing_breakdown.evaluation_run_duration = task_breakdown.evaluation_run_duration
-                    existing_breakdown.sandbox_run_duration = task_breakdown.sandbox_run_duration
-                    commit_task_status_transition(task_row.id, task_session, org, TaskStatus.FINISHED)
-
-                    return {task_id: evaluation_result_row.result}
-            except Exception:
-                with Session(bind=engine) as task_session:
-                    task = fetch_task_row(task_row.id, task_session, org)
-                    if task.status == TaskStatus.STOPPED:
-                        return {task_id: None}
-
-                raise
-
-    except SandboxSetupError as e:
-        if task_is_stopped():
-            return {task_id: None}
-        log_output(f"\n[ERROR] {e}")
-        raise
-    except OutputArtifactError as e:
-        if task_is_stopped():
-            return {task_id: None}
-        error_message = str(e)
-        logger.warning(error_message)
-        log_output(f"\n[ERROR] {error_message}")
-
-        with Session(bind=engine) as task_session:
-            task = fetch_task_row(task_row.id, task_session, org)
-            commit_task_error(task, task_session, error_message)
-
-        return {task_id: None}
-    except ConnectionClosedError:
-        if task_is_stopped():
-            return {task_id: None}
-        seconds = int(time.monotonic() - last_log_time)
-        error_message = (
-            f"Benchmark service has not sent a message, causing the connection to disconnect: "
-            f"last message received {seconds}s ago"
-        )
-        logger.warning(error_message)
-        log_output(f"\n[ERROR] {error_message}")
-
-        with Session(bind=engine) as task_session:
-            task = fetch_task_row(task_row.id, task_session, org)
-            commit_task_error(task, task_session, error_message)
-
-        return {task_id: None}
-    except ValidationError as e:
-        if task_is_stopped():
-            return {task_id: None}
-        field_names = ", ".join(".".join(str(loc) for loc in err["loc"]) for err in e.errors())
-        error_message = (
-            f"Benchmark service returned an incompatible task response. Missing or invalid fields: {field_names}"
-        )
-        log_output(f"\n[ERROR] {error_message}")
-
-        with Session(bind=engine) as task_session:
-            task = fetch_task_row(task_row.id, task_session, org)
-            commit_task_error(task, task_session, error_message)
-
-        return {task_id: None}
-    except InvalidStatus as e:
-        if task_is_stopped():
-            return {task_id: None}
-        error_message = f"Benchmark service rejected the WebSocket connection (HTTP {e.response.status_code})"
-        log_output(f"\n[ERROR] {error_message}")
-
-        with Session(bind=engine) as task_session:
-            task = fetch_task_row(task_row.id, task_session, org)
-            commit_task_error(task, task_session, error_message)
-
-        return {task_id: None}
-    except BenchmarkServiceError as e:
-        if task_is_stopped():
-            return {task_id: None}
-        error_message = str(e)
-        log_output(f"\n[ERROR] {error_message}")
-
-        with Session(bind=engine) as task_session:
-            task = fetch_task_row(task_row.id, task_session, org)
-            commit_task_error(task, task_session, error_message)
-
-        return {task_id: None}
-    except Exception as e:
-        if task_is_stopped():
-            return {task_id: None}
-        logfire.exception("process_task failed")
-        error_message = str(e)
-        logger.error(error_message, exc_info=True)
-
-        sentry_sdk.capture_exception(e)
-
-        # include the error message
-        log_output(f"\n[ERROR] {error_message}")
-
-        with Session(bind=engine) as task_session:
-            task = fetch_task_row(task_row.id, task_session, org)
-            commit_task_error(task, task_session, error_message)
-
-        return {task_id: None}
-    finally:
-        flush_task.cancel()
-        with suppress(asyncio.CancelledError):
-            await flush_task
-        buffer_logs(log_queue, stream_key, harness_config.aws, harness_config.log_group, force_flush=True)
-
-
-def set_benchmark_final_status(benchmark_row: Benchmark, session: Session, org: Org) -> None:
-    """
-    Delegates status depending on if any tasks have been stopped.
-    """
-
-    # Check if any tasks are still in the pending or in progress state
-    tasks_not_finished: int = session.exec(
-        select(func.count(col(Task.id)))
-        .where(col(Task.benchmark) == benchmark_row.id)
-        .where(col(Task.org_id) == org.id)
-        .where(col(Task.status).in_(_RUNNABLE_TASK_STATUSES))
-    ).one()
-
-    # Tasks will be in a non-finished state if something interrupts them while they are running and the state errors here
-    if tasks_not_finished:
-        raise TrackerServiceError(
-            f"Cannot set final status for run {benchmark_row.id} because tasks are still runnable."
-        )
-
-    tasks_stopped: int = session.exec(
-        select(func.count(col(Task.id)))
-        .where(col(Task.benchmark) == benchmark_row.id)
-        .where(col(Task.org_id) == org.id)
-        .where(col(Task.status) == TaskStatus.STOPPED)
-    ).one()
-
-    # Default status is finished, if we stopped any tasks the benchmark status is stopped
-    # Later we can use the stopped status to determine if we can resume the benchmark
-    benchmark_status = BenchmarkStatus.FINISHED
-    if tasks_stopped:
-        benchmark_status = BenchmarkStatus.STOPPED
-
-    benchmark_row.status = benchmark_status
-    benchmark_row.error_message = None
-    session.add(benchmark_row)
-    session.commit()
-
-
-def create_task_rows(
-    verified_task_ids: list[str],
-    benchmark_row: Benchmark,
-    session: Session,
-    org: Org,
-) -> Sequence[tuple[str, Task]]:
-    """
-    Create task_rows that do not already exist in the database for the benchmark row.
-
-    NOTE: Only return runnable tasks to support resuming the benchmark.
-    """
-    # Find task ids that already exist so that we can filter them out
-    existing_task_ids: Sequence[str] = session.exec(
-        select(Task.task_id).where(Task.benchmark == benchmark_row.id).where(col(Task.task_id).in_(verified_task_ids))
-    ).all()
-
-    # NOTE: Must maintain same order that was passed in
-    task_ids_to_create = [task_id for task_id in verified_task_ids if task_id not in existing_task_ids]
-
-    for task_id in task_ids_to_create:
-        task_row = Task(org_id=org.id, task_id=task_id, benchmark=benchmark_row.id)
-        session.add(task_row)
-
-    session.commit()
-    session.expire_all()
-
-    task_rows = session.exec(
-        select(Task.task_id, Task)
-        .where(Task.benchmark == benchmark_row.id)
-        .where(Task.org_id == org.id)
-        .where(col(Task.task_id).in_(verified_task_ids))
-        .where(col(Task.status).in_([TaskStatus.PENDING, TaskStatus.EVALUATING]))
-    ).all()
-
-    task_rows_by_id: dict[str, Task] = {task_id: task_row for task_id, task_row in task_rows}
-    return [(task_id, task_rows_by_id[task_id]) for task_id in verified_task_ids if task_id in task_rows_by_id]
-
-
-def has_runnable_tasks(session: Session, benchmark_row: Benchmark, org: Org) -> bool:
-    return (
-        session.exec(
-            select(Task.id)
-            .where(Task.benchmark == benchmark_row.id)
-            .where(Task.org_id == org.id)
-            .where(col(Task.status).in_(_RUNNABLE_TASK_STATUSES))
-        ).first()
-        is not None
-    )
-
-
-def has_stopped_tasks(session: Session, benchmark_row: Benchmark, org: Org) -> bool:
-    return (
-        session.exec(
-            select(Task.id)
-            .where(Task.benchmark == benchmark_row.id)
-            .where(Task.org_id == org.id)
-            .where(Task.status == TaskStatus.STOPPED)
-        ).first()
-        is not None
-    )
-
-
-def fetch_final_score_inputs(session: Session, benchmark_row: Benchmark, org: Org) -> dict[str, dict[str, Any] | None]:
-    """
-    Return a mapping of the task IDs to their latest evaluation results. If a task is not finished we default to None.
-    """
-    # Fetch task rows which belong to the benchmark we are running
-    task_rows = cast(
-        Sequence[tuple[UUID, str, TaskStatus]],
-        session.exec(
-            select(Task.id, Task.task_id, Task.status)
-            .where(col(Task.benchmark) == benchmark_row.id)
-            .where(col(Task.org_id) == org.id)
-        ).all(),
-    )
-
-    # Fetch all results from tasks that are finished
-    task_row_ids = [task_row_id for task_row_id, _task_id, status in task_rows if status == TaskStatus.FINISHED]
-    result_rows = cast(
-        Sequence[tuple[UUID, dict[str, Any]]],
-        session.exec(
-            select(EvaluationResult.task, EvaluationResult.result)  # pyright: ignore[reportUnknownArgumentType]
-            .where(col(EvaluationResult.task).in_(task_row_ids))
-            .where(col(EvaluationResult.org_id) == org.id)
-            .order_by(desc(EvaluationResult.created_at))
-        ).all(),
-    )
-
-    # Group results by task row ID
-    latest_results: dict[UUID, dict[str, Any]] = {}
-    for task_row_id, result in result_rows:
-        latest_results.setdefault(task_row_id, result)
-
-    # Return mapping between task IDs and their latest evaluation results
-    return {
-        task_id: latest_results.get(task_row_id) if status == TaskStatus.FINISHED else None
-        for task_row_id, task_id, status in task_rows
-    }
-
-
-@broker.task
-@logfire.instrument("process_benchmark")
-async def process_benchmark(
-    start_benchmark_request_json: dict[str, Any],
-    benchmark_id_str: str,
-    verified_task_ids: list[str],
-) -> None:
-    # Was serialized to make it compatible with the broker
-    start_benchmark_request: StartBenchmarkRequest = StartBenchmarkRequest(**start_benchmark_request_json)
-    benchmark_id: UUID = UUID(benchmark_id_str)
-    harness_config: HarnessConfig = start_benchmark_request.harness_config
-    sandbox_provider_config = fetch_sandbox_provider_config(
-        harness_config.sandbox_provider_secret_name,
-        harness_config.aws,
-        start_benchmark_request.sandbox_provider,
-    )
-    benchmark_service = create_benchmark_service_client_from_request(start_benchmark_request)
-
-    sentry_sdk.set_tag("benchmark_name", start_benchmark_request.benchmark_name)
-    sentry_sdk.set_tag("agent_name", start_benchmark_request.contract.name)
-    trace.get_current_span().set_attributes(
-        {
-            "benchmark_id": benchmark_id_str,
-            "benchmark_name": start_benchmark_request.benchmark_name,
-            "agent_name": start_benchmark_request.contract.name,
-            "task_count": len(verified_task_ids),
-        }
-    )
-
-    # Create notifier if webhook is configured
-    notifier: SlackNotifier | None = None
-    if start_benchmark_request.webhook_secret_name and start_benchmark_request.webhook_intervals:
-        notifier = SlackNotifier(
-            secret_name=start_benchmark_request.webhook_secret_name,
-            aws=harness_config.aws,
-            intervals=start_benchmark_request.webhook_intervals,
-        )
-
-    # Resolve the org from the benchmark row (no org check on first fetch since the benchmark was just created by our system)
-    with Session(bind=engine) as session:
-        benchmark_row = session.get(Benchmark, benchmark_id)
-        if not benchmark_row:
-            raise TrackerServiceError(f"Run with id {benchmark_id} not found")
-        org = session.exec(select(Org).where(Org.id == benchmark_row.org_id)).one()
-
-    finalization_deferred = False
-    try:
-        # Copy the agent into the benchmarks S3 folder
-        await copy_agent_to_benchmark(
-            str(benchmark_id),
-            start_benchmark_request.contract.name,
-            harness_config.aws,
-            harness_config.s3_bucket,
-        )
-
-        # Create benchmark cloudwatch log group
-        create_benchmark_log_group(
-            str(benchmark_id), harness_config.aws, harness_config.log_group, harness_config.log_retention_policy
-        )
-
-        # Create tasks inside of the database for each task id
-        with Session(bind=engine) as session:
-            benchmark_row = fetch_benchmark_row(benchmark_id, session, org)
-            task_rows: Sequence[tuple[str, Task]] = create_task_rows(verified_task_ids, benchmark_row, session, org)
-
-        task_row_ids: set[str] = {task_id for task_id, _ in task_rows}
-        missing_task_ids: list[str] = [task_id for task_id in verified_task_ids if task_id not in task_row_ids]
-        if missing_task_ids:
-            raise TrackerServiceError(
-                f"Race condition occured when resuming run {benchmark_id}. Missing task ids: {', '.join(missing_task_ids)}"
-            )
-
-        # Semaphore to isolate concurrent sandboxes that are being made for the benchmark
-        creation_semaphore = Semaphore(_SANDBOX_CREATION_CAP)
-
-        # Load the tasks we are going to be tracking
-        tracked_tasks: dict[str, TrackedTask] = {
-            task_id: TrackedTask(
-                process_task(
-                    task_row,
-                    start_benchmark_request,
-                    benchmark_service,
-                    benchmark_id,
-                    task_id,
-                    harness_config,
-                    org,
-                    sandbox_provider_config=sandbox_provider_config,
-                    creation_semaphore=creation_semaphore,
-                ),
-                org,
-            )
-            for task_id, task_row in task_rows
-        }
-
-        # Start the monitor to track the state the tasks are in and cancel them when no longer valid
-        monitor = TaskMonitor(benchmark_id, tracked_tasks, org, notifier=notifier)
-        monitor_task = asyncio.create_task(monitor.track_tasks())
-
-        semaphore = Semaphore(start_benchmark_request.concurrency)
-
-        await gather(*[tracked_tasks[task_id].run(semaphore, task_row) for task_id, task_row in task_rows])
-
-        await monitor_task
-
-        with Session(bind=engine) as session:
-            benchmark_row = fetch_benchmark_row(benchmark_id, session, org)
-            if has_runnable_tasks(session, benchmark_row, org):
-                finalization_deferred = True
-                return
-            evaluation_results = fetch_final_score_inputs(session, benchmark_row, org)
-
-        if not any(result is not None for result in evaluation_results.values()):
-            with Session(bind=engine) as session:
-                benchmark_row = fetch_benchmark_row(benchmark_id, session, org)
-                if has_stopped_tasks(session, benchmark_row, org):
-                    set_benchmark_final_status(benchmark_row, session, org)
-                    return
-            raise TrackerServiceError("No tasks were completed successfully")
-
-        # Calculate the final score based off the tasks that were ran
-        final_score_response = await benchmark_service.final_score(
-            evaluation_results=evaluation_results, dataset=start_benchmark_request.dataset
-        )
-
-        with Session(bind=engine) as session:
-            benchmark_row = fetch_benchmark_row(benchmark_id, session, org)
-            # final_score is a network call; a concurrent retry can make tasks runnable before we write FinalEvaluation.
-            if has_runnable_tasks(session, benchmark_row, org):
-                finalization_deferred = True
-                return
-
-        # Create the final evaluation row and add it to the database
-        final_evaluation_row = FinalEvaluation(
-            org_id=org.id,
-            benchmark=benchmark_id,
-            final_score=final_score_response.final_score,
-            properties=final_score_response.metadata,
-        )
-
-        with Session(bind=engine) as session:
-            benchmark_row = fetch_benchmark_row(benchmark_id, session, org)
-            # Delete existing final evaluation if re-running
-            if benchmark_row.final_evaluation:
-                session.delete(benchmark_row.final_evaluation)
-                session.flush()
-
-            session.add(final_evaluation_row)
-            session.commit()
-
-            set_benchmark_final_status(benchmark_row, session, org)
-
-            # Push the final benchmark view to the bucket
-            final_view: FinalViewResponse = create_final_view(benchmark_row, session, org)
-
-            await upload_final_view(benchmark_row, final_view, harness_config)
-
-            # If the user has chosen to invoke a lambda function at the end of the benchmark
-            # We run it but do not let a failure affect the benchmark status
-            arguments = benchmark_row.arguments
-            if arguments.lambda_function:
-                # Expose the benchmark arguments and the benchmark id inside of the lambda
-                lambda_payload: dict[str, Any] = arguments.model_dump()
-                lambda_payload["benchmark_id"] = str(benchmark_id)
-
-                invoke_lambda(lambda_client(harness_config.aws), arguments.lambda_function, lambda_payload)
-
-    except BenchmarkServiceUnauthenticatedError as e:
-        logfire.warn("process_benchmark failed due to benchmark service auth error")
-        with Session(bind=engine) as session:
-            benchmark_row = fetch_benchmark_row(benchmark_id, session, org)
-            error_message = f"{str(e)}\n{traceback.format_exc()}"
-            logger.warning(error_message)
-            commit_benchmark_error(benchmark_row, session, error_message)
-    except BenchmarkServiceError as e:
-        with Session(bind=engine) as session:
-            benchmark_row = fetch_benchmark_row(benchmark_id, session, org)
-            error_message = str(e)
-            commit_benchmark_error(benchmark_row, session, error_message)
-    except Exception as e:
-        logfire.exception("process_benchmark failed")
-        sentry_sdk.capture_exception(e)
-        with Session(bind=engine) as session:
-            benchmark_row = fetch_benchmark_row(benchmark_id, session, org)
-            error_message = f"{str(e)}\n{traceback.format_exc()}"
-            commit_benchmark_error(benchmark_row, session, error_message)
-    finally:
-        if not finalization_deferred:
-            with Session(bind=engine) as session:
-                # Handle any misalignments between the benchmark status and tasks
-                catch_errors_during_cleanup(benchmark_id, session, org)
-
-        if notifier and not finalization_deferred:
-            try:
-                with Session(bind=engine) as session:
-                    benchmark_row = fetch_benchmark_row(benchmark_id, session, org)
-                    notification_context = NotificationContext.from_benchmark(benchmark_row, session, org)
-                    final_score = benchmark_row.final_evaluation.final_score if benchmark_row.final_evaluation else None
-                    await notifier.send_terminal_notification(
-                        notification_context,
-                        status=benchmark_row.status,
-                        final_score=final_score,
-                        error_message=benchmark_row.error_message,
-                    )
-            except Exception as notification_error:
-                logger.warning(f"Failed to send terminal notification: {notification_error}")
-
-        await benchmark_service.close()
-
-
-class TaskCounts(NamedTuple):
-    total_tasks: int
-    finished_tasks: int
-    failed_tasks: int
-
-
-class BenchmarkContext:
-    _benchmark_row: Benchmark
-    _session: Session
-    _org: Org
-
-    def __init__(self, benchmark_row: Benchmark, session: Session, org: Org):
-        self._benchmark_row = benchmark_row
-        self._session = session
-        self._org = org
-
-    @property
-    def _status(self) -> BenchmarkStatus:
-        return self._benchmark_row.status
-
-    @cached_property
-    def _task_counts(self) -> TaskCounts:
-        finished_statuses = [TaskStatus.FINISHED, TaskStatus.ERROR, TaskStatus.STOPPED]
-
-        statement = (
-            select(
-                func.count().label("total_tasks"),
-                func.count(case((col(Task.status).in_(finished_statuses), 1))).label("finished_tasks"),
-                func.count(case((Task.status == TaskStatus.ERROR, 1))).label("failed_tasks"),
-            )
-            .select_from(Task)
-            .where(Task.benchmark == self._benchmark_row.id)
-            .where(Task.org_id == self._org.id)
-        )
-
-        result = self._session.exec(statement).one()
-
-        task_counts = TaskCounts(total_tasks=result[0], finished_tasks=result[1], failed_tasks=result[2])
-
-        return task_counts
-
-    @property
-    def _task_breakdown(self) -> dict[TaskStatus, int]:
-        """
-        Returns a mapping between the task status and the number of tasks in that status
-
-        Provides a breakdown of the benchmark.
-        """
-        statement = (
-            select(Task.status, func.count(col(Task.id)))
-            .select_from(Task)
-            .where(Task.benchmark == self._benchmark_row.id)
-            .where(Task.org_id == self._org.id)
-            .group_by(Task.status)
-            .having(func.count(col(Task.id)) > 0)  # Exclude all with count of 0
-        )
-
-        result = self._session.exec(statement).all()
-
-        if not result:
-            raise TrackerServiceError(
-                f"No tasks have been discovered for run {self._benchmark_row.id}, cannot provide task breakdown"
-            )
-
-        return {TaskStatus(status): count for status, count in result}
-
-    @cached_property
-    def benchmark_details(self) -> BenchmarkDetails:
-        return BenchmarkDetails(
-            status=self._status,
-            started_at=self._benchmark_row.started_at,
-            total_tasks=self._task_counts.total_tasks,
-            finished_tasks=self._task_counts.finished_tasks,
-            task_breakdown=self._task_breakdown,
-            docent_reading_status=self._benchmark_row.docent_reading_status,
-            docent_reading_url=self._benchmark_row.docent_reading_url,
-        )
-
-
-def _fetch_result_histories(
-    session: Session, task_row_ids: Sequence[UUID], current_evaluation_result_ids: set[UUID], org_id: UUID
-) -> dict[UUID, list[dict[str, Any]]]:
-    """
-    Fetch the history of evaluation + error results for the provided task ids, mixing them with the evaluation results we already have.
-    """
-    # Prep query to fetch all evaluation results for the given tasks
-    evaluation_statement = cast(
-        Any,
-        select(EvaluationResult.id, EvaluationResult.task, EvaluationResult.created_at, EvaluationResult.result)
-        .where(col(EvaluationResult.task).in_(task_row_ids))
-        .where(col(EvaluationResult.org_id) == org_id),
-    )
-
-    # Exclude evaluation results we have already collected
-    if current_evaluation_result_ids:
-        evaluation_statement = evaluation_statement.where(
-            col(EvaluationResult.id).notin_(current_evaluation_result_ids)
-        )
-
-    # Fetched evaluation results
-    evaluation_query_result = cast(Any, session.exec(evaluation_statement))
-    evaluation_rows = cast(
-        Sequence[tuple[UUID, UUID, datetime, dict[str, Any]]],
-        evaluation_query_result.all(),
-    )
-
-    # Fetch all of the error results from the provided task_rows (A task can have a error message and a evaluation result depending on if its been reran)
-    error_rows = session.exec(
-        select(ErrorResult.task, ErrorResult.created_at, ErrorResult.error_message)
-        .where(col(ErrorResult.task).in_(task_row_ids))
-        .where(col(ErrorResult.org_id) == org_id)
-    ).all()
-
-    # Create a mapping of the task row, time stamp of when the result for the row was created, and the resulting row
-    # We do this so that we can easily sort it downstream
-    histories: dict[UUID, list[dict[str, Any]]] = {}
-    for _result_id, task_row_id, created_at, result in evaluation_rows:
-        histories.setdefault(task_row_id, []).append(_history_result(created_at, result))
-    for task_row_id, created_at, error_message in error_rows:
-        histories.setdefault(task_row_id, []).append(_history_error(created_at, error_message))
-
-    return {
-        task_row_id: sorted(entries, key=lambda entry: entry["created_at"], reverse=True)
-        for task_row_id, entries in histories.items()
-        if entries
-    }
-
-
-def fetch_evaluation_results(benchmark_id: UUID, session: Session, org_id: UUID) -> dict[str, dict[str, Any]]:
-    """Select the latest successful evaluation result for each finished task."""
-    statement = (
-        select(EvaluationResult, Task.id, Task.task_id, TaskBreakdown)
-        .join(Task, col(EvaluationResult.task) == col(Task.id))
-        .outerjoin(TaskBreakdown, col(Task.task_breakdown) == col(TaskBreakdown.id))
-        .where(Task.benchmark == benchmark_id)
-        .where(Task.org_id == org_id)
-        .where(Task.status == TaskStatus.FINISHED)
-        .order_by(desc(EvaluationResult.created_at))
-    )
-
-    results = cast(
-        Sequence[tuple[EvaluationResult, UUID, str, TaskBreakdown | None]],
-        session.exec(statement).all(),  # pyright: ignore[reportUnknownArgumentType]
-    )
-
-    latest_results: list[tuple[EvaluationResult, UUID, str, TaskBreakdown | None]] = []
-    seen_task_row_ids: set[UUID] = set()
-    for row in results:
-        task_row_id = row[1]
-        if task_row_id not in seen_task_row_ids:
-            latest_results.append(row)
-            seen_task_row_ids.add(task_row_id)
-
-    histories = _fetch_result_histories(
-        session,
-        list(seen_task_row_ids),
-        {evaluation_result.id for evaluation_result, _task_row_id, _task_id, _breakdown in latest_results},
-        org_id,
-    )
-
-    evaluation_results: dict[str, dict[str, Any]] = {}
-    for evaluation_result, task_row_id, task_id, task_breakdown in latest_results:
-        result_data = dict(evaluation_result.result)
-        result_data["agent_caused_exit_reason"] = evaluation_result.agent_caused_exit_reason
-        if task_breakdown is not None:
-            result_data["task_breakdown"] = task_breakdown.model_dump()
-        task_history = histories.get(task_row_id, [])
-        result_data["attempts"] = len(task_history) + 1
-        if task_history:
-            result_data["history"] = task_history
-        evaluation_results[task_id] = result_data
-
-    return evaluation_results
-
-
-def fetch_average_task_breakdown(benchmark_id: UUID, session: Session, org_id: UUID) -> AverageTaskBreakdown | None:
-    """
-    Fetch the average task breakdown for a given benchmark.
-
-    Returns None if there are no task metrics available for the benchmark.
-    """
-    row = session.exec(
-        select(
-            func.avg(TaskBreakdown.sandbox_build_duration),
-            func.avg(TaskBreakdown.agent_run_duration),
-            func.avg(TaskBreakdown.evaluation_run_duration),
-            func.avg(TaskBreakdown.sandbox_run_duration),
-        )
-        .join(Task, col(Task.task_breakdown) == col(TaskBreakdown.id))
-        .where(Task.benchmark == benchmark_id)
-        .where(Task.org_id == org_id)
-    ).one()
-
-    if all(v is None for v in row):
-        return None
-
-    return AverageTaskBreakdown(
-        sandbox_build_duration=row[0],
-        agent_run_duration=row[1],
-        evaluation_run_duration=row[2],
-        sandbox_run_duration=row[3],
-    )
-
-
-def commit_benchmark_error(benchmark_row: Benchmark, session: Session, error_message: str) -> None:
-    benchmark_row.status = BenchmarkStatus.ERROR
-    benchmark_row.error_message = error_message
-    session.add(benchmark_row)
-    session.commit()
-
-
-def catch_errors_during_cleanup(benchmark_id: UUID, session: Session, org: Org) -> None:
-    """
-    On task exit we must clean up any edge cases so that it does not affect the users experience.
-    There are sometimes fishy things that may occur with the sandboxes that if not dealt with could become an issue.
-
-    1. Benchmark status must be in a finished state
-    2. All tasks must be in a finished state
-    """
-    terminal_statuses = [BenchmarkStatus.FINISHED, BenchmarkStatus.ERROR, BenchmarkStatus.STOPPED]
-
-    benchmark_row = fetch_benchmark_row(benchmark_id, session, org)
-    if benchmark_row.status in terminal_statuses:
-        return
-
-    # Force non exited tasks to be ERROR
-    task_terminal_statuses = [TaskStatus.FINISHED, TaskStatus.ERROR, TaskStatus.STOPPED]
-    undetected_exit_tasks = session.exec(
-        select(Task)
-        .where(col(Task.benchmark) == benchmark_id)
-        .where(col(Task.org_id) == org.id)
-        .where(col(Task.status).notin_(task_terminal_statuses))
-    ).all()
-    for task in undetected_exit_tasks:
-        session.add(ErrorResult(org_id=org.id, task=task.id, error_message="Undetected exit of task"))
-        task.status = TaskStatus.ERROR
-        session.add(task)
-    session.commit()
-
-    # Sweep stale RUNNING analyzer invocations to ERROR. The invoke_analyzer
-    # helper uses try/finally so this only fires when the worker process was
-    # killed mid-invocation (no try/finally cleanup ran).
-    if benchmark_row.docent_reading_status == DocentReadingStatus.RUNNING:
-        benchmark_row.docent_reading_status = DocentReadingStatus.ERROR
-        session.add(benchmark_row)
-
-    # Force benchmark to ERROR so that the user knows they can retry any failed tasks
-    commit_benchmark_error(
-        benchmark_row,
-        session,
-        f"Run {benchmark_id} exited without finishing",
-    )
-
-
-def commit_task_error(task_row: Task, session: Session, error_message: str) -> None:
-    session.add(ErrorResult(org_id=task_row.org_id, task=task_row.id, error_message=error_message))
-    _commit_task_status(task_row, session, TaskStatus.ERROR, error_message=error_message)
-
-
-async def stream_benchmark_results(
-    benchmark_id: UUID, session: Session, harness_config: HarnessConfig, org: Org
-) -> AsyncGenerator[str]:
-    """
-    Generate Server-Sent Events with benchmark updates. User connects to this when they want to view live updates of a benchmark.
-
-    Usage from client:
-        curl -X GET http://<endpoint>/stream-benchmark-results/<benchmark_id>?connect=true
-
-    Returns:
-        AsyncGenerator[str]
-    """
-    PULL_INTERVAL = 5
-
-    EVENT_COMPLETE = "event: complete\n\n"
-    EVENT_ERROR = "event: error\ndata:"
-    DATA_PREFIX = "data:"
-    DISCONNECT = "event: disconnect\n\n"
-
-    try:
-        while True:
-            with Session(bind=session.bind) as fresh_session:
-                fresh_benchmark = fresh_session.get(Benchmark, benchmark_id)
-                if not fresh_benchmark or fresh_benchmark.org_id != org.id:
-                    yield f"{EVENT_ERROR} {json.dumps({'error': 'Run not found'})}\n\n"
-                    break
-
-                fresh_session.refresh(fresh_benchmark)
-                benchmark_context = BenchmarkContext(fresh_benchmark, fresh_session, org)
-
-                response_data = FetchBenchmarkResponse(
-                    benchmark_name=fresh_benchmark.name,
-                    benchmark_id=fresh_benchmark.id,
-                    details=benchmark_context.benchmark_details,
-                    s3_bucket_url=create_benchmark_url(
-                        str(fresh_benchmark.id), harness_config.aws.aws_default_region, harness_config.s3_bucket
-                    ),
-                    label=fresh_benchmark.label,
-                    final_score=fresh_benchmark.final_evaluation.final_score
-                    if fresh_benchmark.final_evaluation
-                    else None,
-                )
-
-                yield f"{DATA_PREFIX} {response_data.model_dump_json()}\n\n"
-
-                if fresh_benchmark.status in [BenchmarkStatus.FINISHED, BenchmarkStatus.ERROR, BenchmarkStatus.STOPPED]:
-                    yield EVENT_COMPLETE
-                    break
-
-            await asyncio.sleep(PULL_INTERVAL)
-
-    except asyncio.CancelledError:
-        logger.info(f"Client disconnected from benchmark {benchmark_id} stream")
-        yield DISCONNECT
-
-
-async def initiate_stop_benchmark(benchmark_row: Benchmark, session: Session, force: bool, org: Org) -> None:
-    """
-    Sets the flags to initiate the stopping process for a benchmark.
-
-    Benchmark - Stopping status
-    Tasks - Stopped status
-
-    NOTE: Tasks that have already started will continue to run and finish.
-    """
-    try:
-        # Update all rows where tasks are pending or building to stopped
-        result = session.exec(
-            update(Task)
-            .where(col(Task.benchmark) == benchmark_row.id)
-            .where(col(Task.org_id) == org.id)
-            .where(col(Task.status).in_([TaskStatus.PENDING, TaskStatus.BUILDING, TaskStatus.EVALUATING]))
-            .values(status=TaskStatus.STOPPED)
-        )
-        session.commit()
-
-        # If we have stopped any tasks or are forcing the benchmark to stop, set the benchmark status to stopping
-        if result.rowcount > 0 or force:
-            benchmark_row.status = BenchmarkStatus.STOPPING
-            session.add(benchmark_row)
-            session.commit()
-    except Exception as e:
-        raise TrackerServiceError(f"Unexpected error stopping run {benchmark_row.id}: {str(e)}") from e
-
-
-async def stop_sandbox(sandbox: Sandbox, provider: SandboxProvider) -> str | None:
-    try:
-        await delete_sandbox(sandbox, provider)
-        return None
-    except SandboxNotFoundError:
-        logger.warning(f"Sandbox `{sandbox.name}` has already been terminated")
-        return None
-    except Exception as e:
-        return f"{str(e)}: {traceback.format_exc()}"
-
-
-async def sandbox_generator(benchmark_row: Benchmark, provider: SandboxProvider) -> AsyncGenerator[Sandbox, None]:
-    """
-    Generator that yields all sandboxes for a given benchmark.
-    """
-    query = SandboxQuery(labels={"Benchmark": benchmark_row.name, "Id": str(benchmark_row.id)})
-    async for sandbox in provider.list_sandboxes(query):
-        yield sandbox
-
-
-async def force_stop_sandboxes(
-    benchmark_row: Benchmark,
-    session: Session,
-    sandbox_provider_secret_name: str,
-    aws: AWSCredentials,
-    org: Org,
-    sandbox_provider: str = "daytona",
-) -> None:
-    """
-    Stops and deletes all sandboxes which are in progress or evaluating.
-    NOTE: If task is not in progress but sandbox exists, we kill it and leave the task status as is.
-
-    Raises:
-        TrackerServiceError: If there are any errors stopping the sandboxes
-    """
-    benchmark_service = benchmark_row.benchmark_service()
-    provider = benchmark_service.get_sandbox_provider(
-        fetch_sandbox_provider_config(sandbox_provider_secret_name, aws, sandbox_provider)
-    )
-
-    # Update all tasks being processed to stopped
-    session.exec(
-        update(Task)
-        .where(col(Task.benchmark) == benchmark_row.id)
-        .where(col(Task.org_id) == org.id)
-        .where(col(Task.status).in_([TaskStatus.BUILDING, TaskStatus.IN_PROGRESS, TaskStatus.EVALUATING]))
-        .values(status=TaskStatus.STOPPED)
-    )
-
-    session.commit()
-
-    # Iterate through each running sandbox and stop it, collecting error messages
-    results: dict[str, str | None] = {}
-    try:
-        async for sandbox in sandbox_generator(benchmark_row, provider):
-            result = await stop_sandbox(sandbox, provider)
-            results[sandbox.name] = result
-    finally:
-        await benchmark_service.close()
-
-    error_message: str = "\n".join(
-        f"{task_alias}: {error_message}" for task_alias, error_message in results.items() if error_message
-    )
-
-    # If all tasks are already in a stopped state, we need to update the final status here since the worker has exited
-    finished_statuses: list[TaskStatus] = [TaskStatus.FINISHED, TaskStatus.ERROR, TaskStatus.STOPPED]
-    tasks_still_running: int = session.exec(
-        select(func.count(col(Task.id)))
-        .where(col(Task.benchmark) == benchmark_row.id)
-        .where(col(Task.org_id) == org.id)
-        .where(col(Task.status).notin_(finished_statuses))
-    ).one()
-
-    if not tasks_still_running:
-        benchmark_row.status = BenchmarkStatus.STOPPED
-        session.add(benchmark_row)
-        session.commit()
-
-    if error_message:
-        raise TrackerServiceError(f"Unexpected errors stopping sandboxes:\n{error_message}")
-
-
-async def reset_to_in_progress_status(
-    benchmark_row: Benchmark,
-    session: Session,
-    benchmark_service: BenchmarkServiceClient,
-    retry: bool,
-    retry_mode: RetryMode,
-    rerun_task_ids: list[str],
-    org: Org,
-) -> list[str]:
-    """
-    Resets valid tasks to in progress and to allow for retrying or resuming the benchmark.
-
-    Retry: we reset objects with an error status ontop of the stopped status
-    Rerun Task IDs: even if task has been finished we restart it. If the task has no
-        row yet, a fresh PENDING row is created when valid in the current dataset.
-
-    Benchmark - In progress status
-    Tasks - Pending status, or Evaluating status when retrying durable eval state
-
-    NOTE: Will raise if benchmark is in a stopped state with no stopped tasks.
-    """
-    try:
-        existing_rows = session.exec(
-            select(Task)
-            .where(*_retry_task_filters(benchmark_row, retry, rerun_task_ids, org))
-            .order_by(asc(Task.started_at))
-        ).all()
-        existing_by_task_id: dict[str, Task] = {task.task_id: task for task in existing_rows}
-
-        if benchmark_row.status == BenchmarkStatus.IN_PROGRESS:
-            missing_task_ids = [task_id for task_id in rerun_task_ids if task_id not in existing_by_task_id]
-            if missing_task_ids:
-                raise TrackerServiceError(
-                    f"{', '.join(missing_task_ids)} cannot be retried while run {benchmark_row.id} is in progress because they are not in ERROR status"
-                )
-            new_task_ids = []
-        else:
-            new_task_ids = [tid for tid in rerun_task_ids if tid not in existing_by_task_id]
-
-        # Allow re-running the end of the benchmark without running any tasks
-        if not existing_rows and not new_task_ids:
-            return []
-
-        # Verify the task ids are still valid before priming to resume
-        # Raises if any task ids are invalid
-        all_requested_task_ids = [task.task_id for task in existing_rows] + new_task_ids
-        verify_response = await benchmark_service.verify_task_ids(
-            task_ids=all_requested_task_ids, slice_str=None, dataset=benchmark_row.arguments.dataset
-        )
-
-        # Can already be in progress when retrying errored tasks while the run is ongoing.
-        if benchmark_row.status != BenchmarkStatus.IN_PROGRESS:
-            benchmark_row.status = BenchmarkStatus.IN_PROGRESS
-            session.add(benchmark_row)
-            session.commit()
-
-        for task in existing_rows:
-            task.status = (
-                TaskStatus.EVALUATING
-                if retry_mode == RetryMode.AUTO and task.eval_resume_state is not None
-                else TaskStatus.PENDING
-            )
-            task.started_at = datetime.now(ZoneInfo("UTC"))
-            task.finished_at = None
-            if retry_mode == RetryMode.FROM_SCRATCH:
-                task.eval_resume_state = None
-            session.add(task)
-
-        for task_id in new_task_ids:
-            session.add(Task(org_id=org.id, task_id=task_id, benchmark=benchmark_row.id, status=TaskStatus.PENDING))
-
-        session.commit()
-
-        return verify_response.task_ids
-    except (TrackerServiceError, BenchmarkServiceError):
-        raise
-    except Exception as e:
-        raise TrackerServiceError(f"Unexpected error resuming run {benchmark_row.id}: {str(e)}") from e
-
-
-def encode_cursor(started_at: datetime, row_id: UUID) -> str:
-    """Encode a keyset pagination cursor from a started_at timestamp and row id."""
-    payload = json.dumps({"started_at": started_at.isoformat(), "id": str(row_id)})
-    return base64.urlsafe_b64encode(payload.encode()).decode().rstrip("=")
-
-
-def decode_cursor(cursor: str) -> tuple[datetime, UUID]:
-    """Decode a keyset pagination cursor into a started_at timestamp and row id."""
-    padded = cursor + "=" * (-len(cursor) % 4)
-    payload = json.loads(base64.urlsafe_b64decode(padded.encode()).decode())
-    return datetime.fromisoformat(payload["started_at"]), UUID(payload["id"])
-
-
-def _retry_task_filters(benchmark_row: Benchmark, retry: bool, rerun_task_ids: list[str], org: Org) -> list[Any]:
-    """Select retryable rows.
-
-    Active retries on in-progress runs are limited to ERROR tasks. Finished tasks must wait until the run is terminal.
-    """
-    filters = [
-        col(Task.benchmark) == benchmark_row.id,
-        col(Task.org_id) == org.id,
-    ]
-    if benchmark_row.status == BenchmarkStatus.IN_PROGRESS:
-        filters.append(col(Task.status) == TaskStatus.ERROR)
-        if rerun_task_ids:
-            filters.append(col(Task.task_id).in_(rerun_task_ids))
-        return filters
-
-    if retry and rerun_task_ids:
-        filters.append(col(Task.task_id).in_(rerun_task_ids))
-        return filters
-
-    retry_statuses = [TaskStatus.STOPPED]
-    if retry:
-        retry_statuses.append(TaskStatus.ERROR)
-
-    filters.append(or_(col(Task.status).in_(retry_statuses), col(Task.task_id).in_(rerun_task_ids)))
-    return filters
-
-
-def fetch_filtered_benchmark_rows(
-    request: FetchBenchmarksRequest, session: Session, org: Org
-) -> tuple[Sequence[Benchmark], int | None, str | None]:
-    """
-    Creates a query to fetch benchmark rows from the database based on the fetch benchmark request.
-
-    When request.cursor is a non-empty string, uses keyset pagination (tuple comparison on
-    started_at, id) and returns next_cursor. total_count is None in this path.
-
-    When request.cursor is None or empty, uses legacy offset/limit pagination and returns
-    total_count. next_cursor is None in this path.
-
-    Args:
-        request: FetchBenchmarksRequest
-
-    Returns:
-        tuple[Sequence[Benchmark], int | None, str | None]
-        Sequence of benchmark rows, optional total count, optional next cursor
-
-    """
-
-    query = scoped_select(Benchmark, org).options(selectinload(Benchmark.final_evaluation))
-
-    arguments_json = type_coerce(col(Benchmark.arguments), JSON)
-
-    if request.agent_name:
-        if len(request.agent_name) == 1:
-            query = query.where(arguments_json["contract"]["name"].as_string() == request.agent_name[0])
-        else:
-            query = query.where(col(arguments_json["contract"]["name"].as_string()).in_(request.agent_name))
-
-    if request.model:
-        query = query.where(arguments_json["contract"]["model"].as_string() == request.model)
-
-    if request.dataset:
-        dataset_value = arguments_json["dataset"].as_string()
-        if request.dataset == "default":
-            query = query.where(or_(dataset_value == "default", dataset_value.is_(None)))
-        else:
-            query = query.where(dataset_value == request.dataset)
-
-    if request.label is not None:
-        query = query.where(func.lower(Benchmark.label) == request.label.lower())
-
-    if request.benchmark_name:
-        if len(request.benchmark_name) == 1:
-            query = query.where(Benchmark.name == request.benchmark_name[0])
-        else:
-            query = query.where(col(Benchmark.name).in_(request.benchmark_name))
-
-    if request.status:
-        if len(request.status) == 1:
-            query = query.where(Benchmark.status == request.status[0])
-        else:
-            query = query.where(col(Benchmark.status).in_(request.status))
-
-    if request.started_after is not None:
-        query = query.where(Benchmark.started_at > request.started_after)
-
-    if request.started_before is not None:
-        query = query.where(Benchmark.started_at < request.started_before)
-
-    if request.started_by:
-        normalized_emails = [s.strip().lower() for s in request.started_by if s and s.strip()]
-        if normalized_emails:
-            query = query.where(col(Benchmark.started_by_email).in_(normalized_emails))
-
-    if request.order_by == Order.DESC:
-        query = query.order_by(desc(Benchmark.started_at), desc(Benchmark.id))
-    else:
-        query = query.order_by(asc(Benchmark.started_at), asc(Benchmark.id))
-
-    # Keyset cursor path — skip offset/limit and total_count computation.
-    # cursor="" means first page of keyset mode; non-empty cursor means subsequent page.
-    if request.cursor is not None:
-        if request.cursor:
-            cursor_started_at, cursor_id = decode_cursor(request.cursor)
-
-            if request.order_by == Order.DESC:
-                query = query.where(
-                    tuple_(col(Benchmark.started_at), col(Benchmark.id))
-                    < tuple_(literal(cursor_started_at), literal(str(cursor_id)))
-                )
-            else:
-                query = query.where(
-                    tuple_(col(Benchmark.started_at), col(Benchmark.id))
-                    > tuple_(literal(cursor_started_at), literal(str(cursor_id)))
-                )
-
-        # Fetch one extra row to detect whether there is a next page
-        query = query.limit(request.limit + 1)
-        benchmark_rows: Sequence[Benchmark] = session.exec(query).all()
-
-        next_cursor: str | None = None
-        if len(benchmark_rows) > request.limit:
-            benchmark_rows = benchmark_rows[: request.limit]
-            last_row = benchmark_rows[-1]
-            next_cursor = encode_cursor(last_row.started_at, last_row.id)
-
-        return benchmark_rows, None, next_cursor
-
-    # Legacy offset/limit path — compute total_count for backward compat
-    total_count = session.exec(select(func.count()).select_from(query.subquery())).one()
-
-    if not total_count:
-        return [], 0, None
-
-    query = query.limit(request.limit).offset(request.offset)
-    benchmark_rows = session.exec(query).all()
-
-    return benchmark_rows, total_count, None
-
-
-def build_benchmark_table_rows(benchmarks: Sequence[Benchmark], session: Session) -> list[BenchmarkTableRow]:
-    """Batch-load task counts + run-by emails for a page of benchmarks.
-
-    Caller must have eager-loaded `final_evaluation`. Avoids the N+1 of
-    Benchmark.create_benchmark_table_row in a loop.
-    """
-    if not benchmarks:
-        return []
-
-    bench_ids = [b.id for b in benchmarks]
-
-    count_rows = session.exec(
-        select(Task.benchmark, Task.status, func.count())
-        .where(col(Task.benchmark).in_(bench_ids))
-        .group_by(col(Task.benchmark), col(Task.status))
-    ).all()
-    counts_by_bench: dict[UUID, dict[TaskStatus, int]] = {}
-    for bench_id, status, count in count_rows:
-        counts_by_bench.setdefault(bench_id, {})[TaskStatus(status)] = count
-
-    rows: list[BenchmarkTableRow] = []
-    for b in benchmarks:
-        counts = counts_by_bench.get(b.id, {})
-        rows.append(
-            BenchmarkTableRow(
-                id=b.id,
-                name=b.name,
-                agent_name=b.arguments.contract.name,
-                label=b.label,
-                model=b.arguments.contract.model,
-                dataset=b.arguments.dataset or "default",
-                started_by_email=b.started_by_email,
-                started_at=b.started_at,
-                finished_at=b.finished_at,
-                status=b.status,
-                total_tasks=sum(counts.values()),
-                finished_tasks=(
-                    counts.get(TaskStatus.FINISHED, 0)
-                    + counts.get(TaskStatus.ERROR, 0)
-                    + counts.get(TaskStatus.STOPPED, 0)
-                ),
-                task_state_counts={k.value: v for k, v in counts.items()},
-                final_score=b.final_evaluation.final_score if b.final_evaluation else None,
-            )
-        )
-    return rows
-
-
-class YieldingWriter(io.RawIOBase):
-    """
-    Custom writer that collects bytes and returns them to stream.
-    """
-
-    def __init__(self):
-        super().__init__()
-        self._buffer: bytearray = bytearray()
-
-    def writable(self) -> bool:
-        return True
-
-    def write(self, b: Buffer) -> int:
-        data = bytes(b)
-        self._buffer.extend(data)
-
-        return len(data)
-
-    def pop(self) -> bytes:
-        if not self._buffer:
-            return b""
-
-        chunk = bytes(self._buffer)
-        self._buffer.clear()
-
-        return chunk
-
-
-def create_final_view(benchmark_row: Benchmark, session: Session, org: Org) -> FinalViewResponse:
-    """Creates final view of a benchmark that includes metadata about evaluations and score"""
-    tasks_stopped: int = session.exec(
-        select(func.count(col(Task.id)))
-        .where(col(Task.benchmark) == benchmark_row.id)
-        .where(col(Task.org_id) == org.id)
-        .where(col(Task.status) == TaskStatus.STOPPED)
-    ).one()
-
-    final_view: FinalViewResponse = FinalViewResponse(
-        benchmark_name=benchmark_row.name,
-        status=benchmark_row.status,
-        error_message=benchmark_row.error_message,
-        benchmark_id=benchmark_row.id,
-        benchmark_arguments=benchmark_row.arguments,
-        started_at=benchmark_row.started_at,
-        finished_at=benchmark_row.finished_at,
-        tasks_stopped=tasks_stopped or None,  # NOTE: Only include if we stopped the benchmark
-        final_evaluation=benchmark_row.final_evaluation,
-        evaluation_results=benchmark_row.fetch_evaluation_results(session),
-        task_errors=benchmark_row.fetch_tasks_with_errors(session),
-        average_task_breakdown=fetch_average_task_breakdown(benchmark_row.id, session, org.id),
-    )
-
-    return final_view
-
-
-async def upload_final_view(
-    benchmark_row: Benchmark, final_view: FinalViewResponse, harness_config: HarnessConfig
-) -> str:
-    """Uploads the final view to the root of the benchmark folder and returns the s3 key"""
-    s3_key = f"{S3_BENCHMARKS_PREFIX}/{benchmark_row.id}/{benchmark_row.name}.json"
-    await upload_to_s3(
-        final_view.model_dump_json(indent=4, exclude_none=True).encode(),
-        s3_key,
-        harness_config.aws,
-        harness_config.s3_bucket,
-    )
-
-    return s3_key
-
-
-def _parse_log_retention_policy(value: int | str | None, *, source: str) -> int:
-    if value in (None, ""):
-        return 30
-    try:
-        parsed = int(value)
-    except (TypeError, ValueError) as e:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Invalid log_retention_policy from {source}: must be an integer",
-        ) from e
-    if parsed <= 0:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Invalid log_retention_policy from {source}: must be positive",
-        )
-    return parsed
-
-
-_REQUIRED_HARNESS_HEADER_KEYS = ("aws_access_key_id", "aws_secret_access_key", "aws_default_region", "s3_bucket")
-
-
-def _parse_harness_headers(request: Request) -> dict[str, str]:
-    prefix = "x-harness-"
-    return {
-        key[len(prefix) :].replace("-", "_"): value for key, value in request.headers.items() if key.startswith(prefix)
-    }
-
-
-def _build_harness_config(flat: dict[str, str]) -> HarnessConfig:
-    return HarnessConfig(
-        aws=AWSCredentials(
-            aws_access_key_id=flat["aws_access_key_id"],
-            aws_secret_access_key=flat["aws_secret_access_key"],
-            aws_default_region=flat["aws_default_region"],
-            aws_session_token=flat.get("aws_session_token"),
-        ),
-        s3_bucket=flat["s3_bucket"],
-        log_group=flat.get("log_group") or "",
-        log_retention_policy=_parse_log_retention_policy(
-            flat.get("log_retention_policy"),
-            source="request headers",
-        ),
-        sandbox_provider_secret_name=flat.get("sandbox_provider_secret_name") or flat.get("daytona_secret_name") or "",
-    )
-
-
-def try_fetch_harness_config(request: Request) -> HarnessConfig | None:
-    """HarnessConfig from X-Harness-* headers, or None if any required header is missing.
-
-    Used by endpoints (e.g. /start-benchmark) that accept harness_config either from headers (web FE)
-    or from the request body (CLI).
-    """
-    flat = _parse_harness_headers(request)
-    if any(not flat.get(key) for key in _REQUIRED_HARNESS_HEADER_KEYS):
-        return None
-
-    return _build_harness_config(flat)
-
-
-def fetch_harness_config(request: Request) -> HarnessConfig:
-    """HarnessConfig from X-Harness-* headers; 400 naming the first missing required header."""
-    flat = _parse_harness_headers(request)
-    for key in _REQUIRED_HARNESS_HEADER_KEYS:
-        if not flat.get(key):
-            header_name = key.replace("_", "-")
-            raise HTTPException(status_code=400, detail=f"Missing harness config header 'x-harness-{header_name}'")
-
-    return _build_harness_config(flat)
