+++ services/tracker/src/tracker/utils/__init__.py
@@ -1,0 +1,63 @@
+"""Back-compat re-export shim for the former tracker/utils.py.
+
+Everything now lives in focused submodules; these re-exports keep
+`from tracker.utils import X` working for all existing callers.
+"""
+
+from tracker.utils.harness_config import (  # noqa: F401
+    _build_harness_config,
+    _parse_harness_headers,
+    _parse_log_retention_policy,
+    fetch_harness_config,
+    try_fetch_harness_config,
+)
+from tracker.utils.orchestration import (  # noqa: F401
+    catch_errors_during_cleanup,
+    commit_benchmark_error,
+    create_task_rows,
+    fetch_final_score_inputs,
+    has_runnable_tasks,
+    has_stopped_tasks,
+    process_benchmark,
+    set_benchmark_final_status,
+)
+from tracker.utils.reporting import (  # noqa: F401
+    BenchmarkContext,
+    TaskCounts,
+    YieldingWriter,
+    build_benchmark_table_rows,
+    create_final_view,
+    decode_cursor,
+    encode_cursor,
+    fetch_average_task_breakdown,
+    fetch_evaluation_results,
+    fetch_filtered_benchmark_rows,
+    stream_benchmark_results,
+    upload_final_view,
+)
+from tracker.utils.resources import (  # noqa: F401
+    create_benchmark_service_client,
+    create_benchmark_service_client_from_request,
+    fetch_benchmark_row,
+    fetch_sandbox_provider_config,
+    fetch_task_row,
+    start_benchmark_request_to_benchmark,
+)
+from tracker.utils.run_control import (  # noqa: F401
+    force_stop_sandboxes,
+    initiate_stop_benchmark,
+    reset_to_in_progress_status,
+    sandbox_generator,
+    stop_sandbox,
+)
+from tracker.utils.task_execution import (  # noqa: F401
+    TaskMonitor,
+    TrackedTask,
+    TrackedTaskStatus,
+    buffer_logs,
+    commit_task_error,
+    commit_task_status_transition,
+    handle_early_exit,
+    process_task,
+    save_eval_resume_state,
+)
