+++ services/tracker/tests/unit/test_tracker.py
@@ -64,7 +64,8 @@ class TestTracker:
             for task_id in tasks_to_track
         }
 
-        monkeypatch.setattr("tracker.utils.engine", database_session.bind)
+        monkeypatch.setattr("tracker.utils.task_execution.engine", database_session.bind)
+        monkeypatch.setattr("tracker.utils.orchestration.engine", database_session.bind)
         monitor = TaskMonitor(benchmark_row.id, task_tracking.copy(), org=self._test_org)
 
         # Change task status to running and add a task to the object
